+++ packages/core/src/domain/telemetry/telemetryEvent.types.ts
@@ -443,9 +443,21 @@ export type TelemetryCommonFeaturesUsage =
   | GetViewContext
   | AddAction
   | AddError
+  | GetGlobalContext
   | SetGlobalContext
+  | SetGlobalContextProperty
+  | RemoveGlobalContextProperty
+  | ClearGlobalContext
+  | GetUser
   | SetUser
+  | SetUserProperty
+  | RemoveUserProperty
+  | ClearUser
+  | GetAccount
   | SetAccount
+  | SetAccountProperty
+  | RemoveAccountProperty
+  | ClearAccount
   | AddFeatureFlagEvaluation
 /**
  * Schema of browser specific features usage
@@ -650,18 +662,81 @@ export interface AddError {
   feature: 'add-error'
   [k: string]: unknown
 }
+export interface GetGlobalContext {
+  /**
+   * getGlobalContext API
+   */
+  feature: 'get-global-context'
+  [k: string]: unknown
+}
 export interface SetGlobalContext {
   /**
-   * setGlobalContext, setGlobalContextProperty, addAttribute APIs
+   * setGlobalContext, addAttribute APIs
    */
   feature: 'set-global-context'
   [k: string]: unknown
 }
+export interface SetGlobalContextProperty {
+  /**
+   * setGlobalContextProperty API
+   */
+  feature: 'set-global-context-property'
+  [k: string]: unknown
+}
+export interface RemoveGlobalContextProperty {
+  /**
+   * removeGlobalContextProperty API
+   */
+  feature: 'remove-global-context-property'
+  [k: string]: unknown
+}
+export interface ClearGlobalContext {
+  /**
+   * clearGlobalContext API
+   */
+  feature: 'clear-global-context'
+  [k: string]: unknown
+}
+export interface GetUser {
+  /**
+   * getUser API
+   */
+  feature: 'get-user'
+  [k: string]: unknown
+}
 export interface SetUser {
   /**
-   * setUser, setUserProperty, setUserInfo APIs
+   * setUser, setUserInfo APIs
    */
   feature: 'set-user'
+  [k: string]: unknown
+}
+export interface SetUserProperty {
+  /**
+   * setUserProperty API
+   */
+  feature: 'set-user-property'
+  [k: string]: unknown
+}
+export interface RemoveUserProperty {
+  /**
+   * removeUserProperty API
+   */
+  feature: 'remove-user-property'
+  [k: string]: unknown
+}
+export interface ClearUser {
+  /**
+   * clearUser API
+   */
+  feature: 'clear-user'
+  [k: string]: unknown
+}
+export interface GetAccount {
+  /**
+   * getAccount API
+   */
+  feature: 'get-account'
   [k: string]: unknown
 }
 export interface SetAccount {
@@ -669,6 +744,27 @@ export interface SetAccount {
    * setAccount, setAccountProperty APIs
    */
   feature: 'set-account'
+  [k: string]: unknown
+}
+export interface SetAccountProperty {
+  /**
+   * setAccountProperty API
+   */
+  feature: 'set-account-property'
+  [k: string]: unknown
+}
+export interface RemoveAccountProperty {
+  /**
+   * removeAccountProperty API
+   */
+  feature: 'remove-account-property'
+  [k: string]: unknown
+}
+export interface ClearAccount {
+  /**
+   * clearAccount API
+   */
+  feature: 'clear-account'
   [k: string]: unknown
 }
 export interface AddFeatureFlagEvaluation {