+++ packages/rum-core/src/domain/contexts/userContext.ts
@@ -1,0 +1,22 @@
+import { createContextManager, CustomerDataType, storeContextManager } from '@datadog/browser-core'
+import type { RumConfiguration } from '../configuration'
+
+export function startUserContext(configuration: RumConfiguration) {
+  const userContextManager = buildUserContextManager()
+
+  if (configuration.storeContextsAcrossPages) {
+    storeContextManager(configuration, userContextManager, 'rum', CustomerDataType.User)
+  }
+
+  return userContextManager
+}
+
+export function buildUserContextManager() {
+  return createContextManager('user', {
+    propertiesConfig: {
+      id: { type: 'string' },
+      name: { type: 'string' },
+      email: { type: 'string' },
+    },
+  })
+}