+++ test/e2e/lib/framework/pageSetups.ts
@@ -119,6 +119,7 @@ export function bundleSetup(options: SetupOptions, servers: Servers) {
 
 export function npmSetup(options: SetupOptions, servers: Servers) {
   let header = options.head || ''
+  const body = options.body || ''
 
   if (options.eventBridge) {
     header += setupEventBridge(servers)
@@ -150,7 +151,32 @@ export function npmSetup(options: SetupOptions, servers: Servers) {
 
   return basePage({
     header,
-    body: options.body,
+    body,
+  })
+}
+
+export function reactSetup(options: SetupOptions, servers: Servers) {
+  let header = options.head || ''
+  let body = options.body || ''
+
+  if (options.eventBridge) {
+    header += setupEventBridge(servers)
+  }
+
+  if (options.rum) {
+    header += html`
+      <script type="text/javascript">
+        window.RUM_CONFIGURATION = ${formatConfiguration(options.rum, servers)}
+        window.RUM_CONTEXT = ${JSON.stringify(options.context)}
+      </script>
+    `
+  }
+
+  body += html` <script type="text/javascript" src="./react-app.js"></script> `
+
+  return basePage({
+    header,
+    body,
   })
 }